fix: move authConfirmProviders to infrastructure
