+++ src/infrastructure/module.ts
@@ -23,8 +23,8 @@ import {AuthRoleService} from '../domain/services/AuthRoleService';
 import {AuthFilePermissionService} from '../domain/services/AuthFilePermissionService';
 import {AuthUpdateUserOwnPasswordUseCase} from '../usecases/updatePassword/AuthUpdateUserOwnPasswordUseCase';
 import {AuthRevokeUserActiveLoginsUseCase} from '../usecases/revokeUserActiveLogins/AuthRevokeUserActiveLoginsUseCase';
-import authConfirmProviders from '../domain/services/authConfirmProviders';
 import {AuthConfirmProvidersToken, IAuthConfirmProvider} from '../domain/interfaces/IAuthConfirmProvider';
+import authConfirmProviders from './services/authConfirmProviders';
 import {SessionService} from './services/SessionService';
 import {AuthLoginRepository} from './repositories/AuthLoginRepository';
 import {AuthPermissionRepository} from './repositories/AuthPermissionRepository';
