+++ src/infrastructure/services/authConfirmProviders/index.ts
